refactor: 보드 팩토리 생성
