+++ src/test/java/chess/domain/board/BoardTest.java
@@ -1,19 +1,15 @@
 package chess.domain.board;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-import chess.domain.piece.Bishop;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
-import java.util.List;
+import chess.domain.piece.*;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
@@ -21,7 +17,7 @@ class BoardTest {
 
     @Test
     void 체스보드는_기물의_위치를_가지고_있다() {
-        final Board board = Board.generate();
+        final Board board = BoardFactory.generate();
 
         assertAll(
                 () -> assertThat(board.findPieceOf(new Square(File.A, Rank.ONE))).containsInstanceOf(Rook.class),
@@ -73,7 +69,7 @@ class BoardTest {
         class 나이트의_경우 {
             @Test
             void 이동경로로_이동할_수_있으면_true_반환한다() {
-                final Board board = Board.generate();
+                final Board board = BoardFactory.generate();
                 //Knight
                 final Square source = new Square(File.B, Rank.ONE);
                 final List<Square> routes = List.of(new Square(File.C, Rank.THREE));
@@ -83,7 +79,7 @@ class BoardTest {
 
             @Test
             void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = Board.generate();
+                final Board board = BoardFactory.generate();
                 //Knight
                 final Square source = new Square(File.G, Rank.ONE);
                 final List<Square> routes = List.of(new Square(File.E, Rank.TWO));
@@ -97,7 +93,7 @@ class BoardTest {
 
             @Test
             void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = Board.generate();
+                final Board board = BoardFactory.generate();
                 //Rook
                 final Square source = new Square(File.A, Rank.ONE);
                 final List<Square> routes = List.of(new Square(File.A, Rank.TWO), new Square(File.A, Rank.THREE));
@@ -111,7 +107,7 @@ class BoardTest {
 
             @Test
             void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = Board.generate();
+                final Board board = BoardFactory.generate();
                 //Bishop
                 final Square source = new Square(File.C, Rank.ONE);
                 final List<Square> routes = List.of(new Square(File.D, Rank.TWO), new Square(File.E, Rank.THREE));
@@ -125,7 +121,7 @@ class BoardTest {
 
             @Test
             void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = Board.generate();
+                final Board board = BoardFactory.generate();
                 //Queen
                 final Square source = new Square(File.D, Rank.ONE);
                 final List<Square> routes = List.of(new Square(File.C, Rank.TWO), new Square(File.B, Rank.THREE));
@@ -139,7 +135,7 @@ class BoardTest {
 
             @Test
             void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = Board.generate();
+                final Board board = BoardFactory.generate();
                 //King
                 final Square source = new Square(File.E, Rank.ONE);
                 final List<Square> routes = List.of(new Square(File.F, Rank.TWO));
@@ -154,7 +150,7 @@ class BoardTest {
 
         @Test
         void 폰이_초기_위치일_때_두칸_갈_수_있다() {
-            final Board board = Board.generate();
+            final Board board = BoardFactory.generate();
             final Square source = new Square(File.A, Rank.TWO);
             final List<Square> routes = List.of(new Square(File.A, Rank.FOUR));
 
@@ -163,7 +159,7 @@ class BoardTest {
 
         @Test
         void 폰이_초기_위치일_때_한칸_갈_수_있다() {
-            final Board board = Board.generate();
+            final Board board = BoardFactory.generate();
             final Square source = new Square(File.B, Rank.TWO);
             final List<Square> routes = List.of(new Square(File.B, Rank.THREE));
 
@@ -172,7 +168,7 @@ class BoardTest {
 
         @Test
         void 폰이_초기_위치일_때_대각선으로_갈_수_없다() {
-            final Board board = Board.generate();
+            final Board board = BoardFactory.generate();
             final Square source = new Square(File.C, Rank.TWO);
             final List<Square> routes = List.of(new Square(File.D, Rank.THREE));
 
@@ -182,7 +178,7 @@ class BoardTest {
 
     @Test
     void 기물을_움직인다() {
-        final Board board = Board.generate();
+        final Board board = BoardFactory.generate();
         final Square source = new Square(File.E, Rank.TWO);
         final Square destination = new Square(File.E, Rank.FOUR);
